Delete broken demo project and unused data API
